+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to semantic versioning.
 
 - Remove `flip`. [fp-ts/function::flip](https://gcanti.github.io/fp-ts/modules/function.ts.html#flip) is now curried.
 - Bump minimum supported fp-ts version to 2.13.1.
+- Rename `is` in `Function` to `isInstanceOf` for readability.
 
 ## 0.15.1 (2022-11-18)